move more ts-specific stuff into src/util
