+++ vsc-extension/src/util/ts-source.ts
@@ -1,0 +1,7 @@
+import ts from "typescript";
+
+export interface SourceFileInfo {
+  sourceFile: ts.SourceFile;
+  functions: ts.FunctionDeclaration[];
+}
+export type SourceFileMap = Record<string, SourceFileInfo>;
